revert moving code into codec.h
